% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-electricalpower.R
\docType{data}
\name{electricalpower}
\alias{electricalpower}
\title{Electrical power consumption during a switch operation}
\format{A data frame with 562 rows and 3 columns:
\describe{
\item{X}{The covariate variable which is the time in that case.}
\item{Y1}{A measurement of the electrical power consumed during a switch
operation.}
\item{Y2}{Another measurement of the electrical power consumed during a
switch operation.}
}}
\usage{
electricalpower
}
\description{
This dataset is based on the French railway switches (or points) which
enable trains to be guided from one track to another at a railway junction.
Each measurement represents the electrical power consumed during a switch
operation.
}
\keyword{datasets}
