% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelLearner.R
\name{EM}
\alias{EM}
\title{EM is used to fit a RHLP model.}
\usage{
EM(modelRHLP, n_tries = 1, max_iter = 1500, threshold = 1e-06,
  verbose = FALSE, verbose_IRLS = FALSE)
}
\arguments{
\item{modelRHLP}{A \link{ModelRHLP} object to be fitted.}

\item{n_tries}{Number of times EM algorithm will be launched.
The solution providing the highest log-likelihood will be returned.

If \code{n_tries} > 1, then for the first pass, parameters are initialized
by uniformly segmenting the data into K segments, and for the next passes,
parameters are initialized by randomly segmenting the data into K contiguous
segments.}

\item{max_iter}{The maximum number of iterations for the EM algorithm.}

\item{threshold}{A numeric value specifying the threshold for the relative
difference of log-likelihood between two steps  of the EM as stopping
criteria.}

\item{verbose}{A logical value indicating whether values of the
log-likelihood should be printed during EM iterations.}

\item{verbose_IRLS}{A logical value indicating whether values of the
criterion optimized by IRLS should be printed at each step of the EM
algorithm.}
}
\value{
EM returns an object of class \link{FittedRHLP}.
}
\description{
EM is used to fit a \link[=ModelRHLP]{RHLP} model. The estimation method is
performed by the Expectation-Maximization algorithm.
}
\details{
EM function is based on the EM algorithm. This function alternates
between a E Step (method of the class \link{StatRHLP}) and a M Step
(method of the class \link{ParamRHLP}) until convergence (until the
absolute difference of log-likelihood between two steps of the EM algorithm
is less than the \code{threshold} parameter).
}
\seealso{
\link{FittedRHLP}, \link{ModelRHLP}, \link{ParamRHLP}, \link{StatRHLP}
}
