% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedRHLP.R
\name{FittedRHLP}
\alias{FittedRHLP}
\title{A Reference Class which represents a fitted RHLP model.}
\description{
FittedRHLP represents a \link[=ModelRHLP]{RHLP} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{modelRHLP}}{A \link{ModelRHLP} object. It contains meta-parameters such as K, p,\dots}

\item{\code{paramRHLP}}{A \link{ParamRHLP} object. It contains the estimated values of the parameters.}

\item{\code{statRHLP}}{A \link{StatRHLP} object. It contains all the statistics associated to the RHLP model.}
}}

\seealso{
\link{ModelRHLP}, \link{ParamRHLP}, \link{StatRHLP}
}
