% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRHLP.R
\docType{class}
\name{ModelRHLP-class}
\alias{ModelRHLP-class}
\alias{ModelRHLP}
\title{A Reference Class which represents a fitted RHLP model.}
\description{
ModelRHLP represents an estimated RHLP model.
}
\section{Fields}{

\describe{
\item{\code{paramRHLP}}{A \link{ParamRHLP} object. It contains the estimated
values of the parameters.}

\item{\code{statRHLP}}{A \link{StatRHLP} object. It contains all the
statistics associated to the RHLP model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "estimatedsignal"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"regressors" = } Polynomial regression components.
      \item \code{"estimatedsignal" = } Estimated signal (Sum of the
      polynomial components weighted by the logistic probabilities.
    }
  }
}
By default, all the above graphs are produced.}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamRHLP}, \link{StatRHLP}
}
