% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRHLP.R
\name{ModelRHLP}
\alias{ModelRHLP}
\title{A Reference Class which represents a RHLP model.}
\description{
ModelRHLP represents a RHLP model. It inherits from \link{FData} class. It
contains all meta-paramters such as \code{K}, the number of segments, \code{p} the
order of the polynomial regression\dots
}
\section{Fields}{

\describe{
\item{\code{K}}{The number of regimes (mixture components).}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{q}}{The dimension of the logistic regression. For the purpose of
segmentation, it must be set to 1.}

\item{\code{variance_type}}{Numeric indicating if the model is homoskedastic
(\code{variance_type} = 1) or heteroskedastic (\code{variance_type} = 2).}
}}

\seealso{
\link{FData}
}
