% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRHLP.R
\docType{class}
\name{ParamRHLP-class}
\alias{ParamRHLP-class}
\alias{ParamRHLP}
\title{A Reference Class which contains parameters of a RHLP model.}
\description{
ParamRHLP contains all the parameters of a RHLP model.
}
\section{Fields}{

\describe{
\item{\code{fData}}{\link{FData} object representing the sample.}

\item{\code{K}}{The number of regimes (mixture components).}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{q}}{The dimension of the logistic regression. For the purpose of
segmentation, it must be set to 1.}

\item{\code{variance_type}}{Numeric indicating if the model is homoskedastic
(\code{variance_type} = 1) or heteroskedastic (\code{variance_type} = 2).}

\item{\code{W}}{Parameters of the logistic process.
\eqn{W = w_{1},\dots,w_{K-1}}{W = (w1,\dots,wK-1)} is a matrix of dimension
\eqn{(q + 1, K - 1)}, with \emph{q} the order of the logistic regression.}

\item{\code{beta}}{Parameters of the polynomial regressions.
\eqn{\beta = (\beta_{1},\dots,\beta_{K})}{\beta = (\beta1,\dots,\betaK)} is
a matrix of dimension \eqn{(p + 1, K)}, with \emph{p} the order of the
polynomial regression.}

\item{\code{sigma2}}{The variances for the \emph{K} regimes. If RHLP model is
homoskedastic (\emph{variance_type} = 1) then sigma2 is a matrix of size
\eqn{(1, 1)}, else if RHLP model is heteroskedastic then sigma2 is a matrix
of size \eqn{(K, 1)}.}
}}

\section{Methods}{

\describe{
\item{\code{initParam(try_algo = 1)}}{Method to initialize parameters \code{W}, \code{beta} and
      \code{sigma2}.

      If try_algo = 1 then \code{W}, \code{beta} and \code{sigma2} are
      initialized by segmenting uniformly into \code{K} contiguous segments
      the response Y. Otherwise, \code{W}, \code{beta} and \code{sigma2} are
      initialized by segmenting randomly into \code{K} segments the response Y.}

\item{\code{MStep(statRHLP, verbose_IRLS)}}{Method used in the EM algorithm to learn the parameters of the RHLP model
based on statistics provided by \code{statRHLP}.}
}}

\seealso{
\link{FData}
}
