% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRHLP.R
\name{ParamRHLP}
\alias{ParamRHLP}
\title{A Reference Class which contains parameters of a RHLP model.}
\description{
ParamRHLP contains all the parameters of a \link[=ModelRHLP]{RHLP} model.
}
\section{Fields}{

\describe{
\item{\code{W}}{Parameters of the logistic process.
\eqn{W = w_{1},\dots,w_{K-1}}{W = (w1,\dots,wK-1)} is a matrix of dimension
\eqn{(q + 1, K - 1)}, with \emph{q} the order of the logistic
regression.}

\item{\code{beta}}{Parameters of the polynomial regressions.
\eqn{\beta = (\beta_{1},\dots,\beta_{K})}{\beta = (\beta1,\dots,\betaK)} is
a matrix of dimension \eqn{(p + 1, K)}, with \emph{p} the
order of the polynomial regression.}

\item{\code{sigma}}{The variances for the \emph{K} regimes. If \link[=ModelRHLP]{RHLP}
model is homoskedastic (\emph{variance_type} = 1) then sigma is a matrix of
size \eqn{(1, 1)}, else if \link[=ModelRHLP]{RHLP} model is heteroskedastic then
sigma is a matrix of size \eqn{(K, 1)}.}
}}

\seealso{
\link{ModelRHLP}
}
