% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRHLP.R
\docType{class}
\name{ModelRHLP-class}
\alias{ModelRHLP-class}
\alias{ModelRHLP}
\title{A Reference Class which represents a fitted RHLP model.}
\description{
ModelRHLP represents a \link[=ModelRHLP]{RHLP} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{paramRHLP}}{A \link{ParamRHLP} object. It contains the estimated values of the parameters.}

\item{\code{statRHLP}}{A \link{StatRHLP} object. It contains all the statistics associated to the RHLP model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "meancurve"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item "regressors"
      \item "meancurve"
    }
  }
}}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamRHLP}, \link{StatRHLP}
}
