% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatRHLP.R
\docType{class}
\name{StatRHLP-class}
\alias{StatRHLP-class}
\alias{StatRHLP}
\title{A Reference Class which contains statistics of a RHLP model.}
\description{
StatRHLP contains all the parameters of a \link[=ParamRHLP]{RHLP} model.
}
\section{Fields}{

\describe{
\item{\code{pi_ik}}{Matrix of size \eqn{(m, K)} representing the probabilities
\eqn{P(zi = k; W) = P(z_{ik} = 1; W)}{P(zi = k; W) = P(z_ik = 1; W)} of the
latent variable \eqn{zi,\ i = 1,\dots,m}{zi, i = 1,\dots,m}.}

\item{\code{z_ik}}{Hard segmentation logical matrix of dimension \eqn{(m, K)}
obtained by the Maximum a posteriori (MAP) rule:
\eqn{z_{ik} = 1 \ \textrm{if} \ z_{ik} = \textrm{arg} \ \textrm{max}_{k} \
P(z_i = k | Y, W, \beta) = tau_ik;\ 0 \ \textrm{otherwise}}{z_ik = 1 if z_ik =
arg max_k P(z_i = k | Y, W, \beta) = tau_ik; 0 otherwise}, \eqn{k = 1,\dots,K}.}

\item{\code{klas}}{Column matrix of the labels issued from \code{z_ik}. Its elements are
\eqn{klas(i) = k}, \eqn{k = 1,\dots,K}.}

\item{\code{Ex}}{Column matrix of dimension \emph{m}. \code{Ex} is the curve expectation
: sum of the polynomial components \eqn{\beta_{k} \times X_{i}}{\betak x X_i}
weighted by the logistic probabilities \code{pi_ik}:
\eqn{Ey(i) = \sum_{k = 1}^{K} pi_ik \times \beta_{k} \times X_{i}}{Ey(i) =
\sum_{k=1}^K pi_ik x \betak x X_i}, \eqn{i = 1,\dots,m}.}

\item{\code{loglik}}{Numeric. Log-likelihood of the RHLP model.}

\item{\code{com_loglik}}{Numeric. Complete log-likelihood of the RHLP model.}

\item{\code{stored_loglik}}{List. Stored values of the log-likelihood at each EM
iteration.}

\item{\code{BIC}}{Numeric. Value of the BIC (Bayesian Information Criterion)
criterion. The formula is \eqn{BIC = log\_lik - nu \times
\textrm{log}(m) / 2}{BIC = loglik - nu x log(m) / 2} with \emph{nu} the
degree of freedom of the RHLP model.}

\item{\code{ICL}}{Numeric. Value of the ICL (Integrated Completed Likelihood)
criterion. The formula is \eqn{ICL = com\_loglik - nu \times
\textrm{log}(m) / 2}{ICL = com_loglik - nu x log(m) / 2} with \emph{nu} the
degree of freedom of the RHLP model.}

\item{\code{AIC}}{Numeric. Value of the AIC (Akaike Information Criterion)
criterion. The formula is \eqn{AIC = log\_lik - nu}{AIC = loglik - nu}.}

\item{\code{cpu_time}}{Numeric. Average executon time of a EM step.}

\item{\code{log_piik_fik}}{Matrix of size \eqn{(m, K)} giving the values of the
logarithm of the joint probability
\eqn{P(Y_{i}, \ zi = k)}{P(Yi, zi = k)}, \eqn{i = 1,\dots,m}.}

\item{\code{log_sum_piik_fik}}{Column matrix of size \emph{m} giving the values of
\eqn{\sum_{k = 1}^{K} \textrm{log} P(Y_{i}, \ zi = k)}{\sum_{k = 1}^{K} log
P(Yi, zi = k)}, \eqn{i = 1,\dots,m}.}

\item{\code{tau_ik}}{Matrix of size \eqn{(m, K)} giving the posterior probability that
\eqn{Y_{i}}{Yi} originates from the \eqn{k}-th regression model
\eqn{P(zi = k | Y, W, \beta)}.}

\item{\code{polynomials}}{Matrix of size \eqn{(m, K)} giving the values of
\eqn{\beta_{k} \times X_{i}}{\betak x X_i}, \eqn{i = 1,\dots,m}.}
}}

\section{Methods}{

\describe{
\item{\code{EStep(paramRHLP)}}{Method used in the EM algorithm to update statistics based on parameters
provided by \code{paramRHLP} (prior and posterior probabilities).}
}}

\seealso{
\link{ParamRHLP}, \link{FData}
}
